backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
PyJWT[crypto]
cryptography
